separate create_lexicon and gibbs sampler
